Add ability to undo/redo bone manipulations in the pose editor
The pose editor undo/redo operations are currently associated with the
global undo/redo context which means that an undo/redo will be performed
even if the pose editor window is not currently active. This may not be
the best way to go about this, perhaps it's best to create a separate
context for the pose editor.